+++ examples/test-highlight.js
@@ -10,7 +10,7 @@ let data = "important data in string";
 function testFunction() {
     // FIXME: Fix this later
     const FIXME = 123;
-
+    [SamChen]
     // Test multiple keywords on same line: TODO and important
     console.log("TODO: important task");
 }